moving analysis toosl into `ir.analysis`
