+++ src/bloqade/emulate/codegen/emulator_ir.py
@@ -1,4 +1,4 @@
-from bloqade.codegen.common.assignment_scan import AssignmentScan
+from bloqade.ir.analysis.assignment_scan import AssignmentScan
 from bloqade.ir.location.base import AtomArrangement, SiteFilling
 from bloqade.ir.visitor.analog_circuit import AnalogCircuitVisitor
 from bloqade.ir.visitor.waveform import WaveformVisitor
