+++ src/bloqade/ir/analysis/assignment_scan.py
