+++ src/bloqade/ir/analysis/scan_variables.py
